1509bf risk m/requirements.txt
# ==========================================
# REQUIREMENTS.TXT - Vortex Trading Bot v2.1
# Исправленная версия для совместимости
# ==========================================

# Основные async библиотеки
aiohttp==3.9.1
aiofiles==23.2.0
websockets==12.0

# Конфигурации
PyYAML==6.0.1
python-dotenv==1.0.0

# Telegram Bot (совместимая версия)
python-telegram-bot==20.7

# Математика и анализ данных
numpy==1.24.3
pandas==2.0.3

# HTTP запросы
requests==2.31.0

# База данных
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
redis==5.0.1

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Криптография
cryptography>=42.0.0

# Работа с временем
python-dateutil==2.8.2
pytz==2023.3

# Утилиты
colorama==0.4.6
typing-extensions==4.8.0

# JSON обработка
orjson==3.9.10

# Логирование
loguru==0.7.2

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1

# Дополнительные утилиты
wheel==0.42.0
setuptools==69.0.3
